feat(acquisition_async_ram_rd_uart.v): add new module acquisition_async_ram_rd_uart
This commit adds a new module called acquisition_async_ram_rd_uart. This module is responsible for asynchronously reading data from a RAM and sending it over UART. It includes a state machine that controls the reading and sending of data. The module supports a configurable read count and includes logic for calculating a checksum.

The module has the following inputs:
- rd_clk: the clock signal for reading data from the RAM
- rst_n: the active-low reset signal
- send_en: a signal to start sending data over UART
- rd_data: the data read from the RAM
- uart_tx_busy: a signal indicating if the UART is busy sending data

The module has the following outputs:
- rd_en: a signal to enable reading data from the RAM
- uart_tx_en: a signal to enable sending data over UART
- uart_tx_data: the data to be sent over UART
- rd_busy: a signal indicating if the RAM is busy reading data

The module includes a state machine with the following states:
- IDLE: the initial state, waiting for the send_en signal to be high
- FETCH: reading data from the RAM
- SEND_ENABLE: waiting for the UART to be ready to send data
- SENDING: waiting for the UART to finish sending data

The module also includes logic for calculating a checksum and updating the send_cnt variable.