chore: add MIT license to the project
Add a new file `LICENSE.md` with the MIT license text to the project. This license grants permission to use, copy, modify, merge, publish, distribute, sublicense, and/or sell the software, subject to certain conditions. The license also includes a warranty disclaimer. This addition ensures that the project is properly licensed and provides clarity on the permitted use of the software.